# perf: improve performance in modifySheet