@Gusted 2021-09-21
@alexanderby Could you have a look at this, it's 2 parts that are in the most utterly important logic of Dark Reader, I don't want to miss anything and cause another regression.