@Gusted 2022-01-05
Their have been some issues with edge cases whereby the usage of .cssText was needed. Closing for now, but in the future we could delay the usage of `.cssText` of the first usage(so the loading of a page is faster).